Move packages component back to main file
